Add a photo and biography to the player model
Looks like "docker run --volume" does not respect the container file
system, which I suppose makes some amount of sense since the docker
command was mounted into the container. So we have to use an absolute
path on the host, and cannot refer to a mounted volume in the container.
That's why we have to add the "ec2-user" directory when mounting the
volume, but not when specifying the environment file.